requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.1
django-filter==23.2
djangorestframework==3.14.0
docutils==0.21.2
drf-nested-routers==0.94.1
idna==3.8
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Pillow==10.0.0
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
sqlparse==0.5.1
stripe==6.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
